Correction faille, ajoute du login et authentification

+++ blog/views.py
@@ -1,8 +1,12 @@
+from calendar import error
+
 from django.shortcuts import render, get_object_or_404, redirect
 from django.utils import timezone
 from .models import Task
 from .forms import TaskForm
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.contrib.admin.views.decorators import staff_member_required
 
 
 
@@ -16,7 +20,11 @@ def task_detail(request, pk):
     return render(request, 'blog/task_detail.html', {'task': task})
 
 
+
 def task_new(request):
+    if not request.user.is_authenticated:
+        return redirect("login")
+
     if request.method == "POST":
         form = TaskForm(request.POST)
         if form.is_valid():
@@ -29,7 +37,11 @@ def task_new(request):
     return render(request, 'blog/task_edit.html', {'form': form})
 
 
+
 def task_edit(request, pk):
+    if not request.user.is_authenticated:
+        return redirect("login")
+
     task = get_object_or_404(Task, pk=pk)
     if request.method == "POST":
         form = TaskForm(request.POST, instance=task)
@@ -43,7 +55,11 @@ def task_edit(request, pk):
     return render(request, 'blog/task_edit.html', {'form': form})
 
 
+
 def task_delete(request, pk):
+    if not request.user.is_authenticated:
+        return redirect("login")
+
     task = get_object_or_404(Task, pk=pk)
     if request.method == "POST":
         task.delete()
@@ -60,9 +76,13 @@ def task_toggle_done(request, pk):
 
 
 
+
 from django.http import JsonResponse
 
 def task_toggle_done_ajax(request, pk) :
+    if not request.user.is_authenticated:
+        return JsonResponse({"error": "Unauthorized"}, status=403)
+
     if request.method == "POST" :
         task = get_object_or_404(Task, pk=pk)
         task.status = not task.status
@@ -72,4 +92,35 @@ def task_toggle_done_ajax(request, pk) :
             "status": task.status,
             "task_id": task.pk
         })
+
+
+
+
+from django.contrib.auth import authenticate, login, logout
+from django.shortcuts import render,redirect
+
+def c_login(request):
+    error = None
+
+    if request.method == "POST":
+        username = request.POST.get("username")
+        password = request.POST.get("password")
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('task_list')
+        else:
+            error = "Nom d'utilisateur ou mot de passe incorrect"
+
+    return render(request, "blog/login.html", {"error": error})
+
+
+def c_logout(request):
+    logout(request)
+    return redirect("task_list")
+
+
+
